fix: correct import paths after backend reorganization
- Fixed translationService.js import paths for logger and Translation model
- Logger now correctly imports from shared/utils/logger
- Translation model correctly imports from local translation/models
- All tests passing (37 test suites, 699 tests)
- Server starts successfully without import errors
- Completes backend subsystem-first reorganization
