+++ backend/src/banking/services/categoryAIService.js
@@ -1,6 +1,6 @@
 const logger = require('../../shared/utils/logger');
 const natural = require('natural');
-const translationService = require('../../shared/services/translationService');
+const translationService = require('../../translation/services/translationService');
 const { enhancedKeywordMatcher } = require('./enhanced-keyword-matching');
 const WordTokenizer = natural.WordTokenizer;
 const PorterStemmer = natural.PorterStemmer;
